smart_money_hunter/requirements.txt
aiohttp==3.9.1
python-telegram-bot==20.7
sqlalchemy==1.4.46
python-dotenv==1.0.0
asyncio
